src/django/app/requirements.txt
behave==v1.2.7.dev6
behave-django
coverage
Django
django-createsuperuser
django-debug-toolbar
django-extensions
django-filter
django-tracer
djangorestframework
djangorestframework-simplejwt
drf-spectacular
./enricher
httpx
json-log-formatter
markdown
psycopg2
pydot
pygraphviz
pyparsing
python-json-logger
pyyaml
redis
requests
sse-starlette
standardwebhooks
starlette
strawberry-django-auth
strawberry-graphql-django
svix
uritemplate
uvicorn[standard]
whitenoise
